feat: Map articles
